Remove the "user defined" example in the Reduction Operators section
As far as I can see, this example does not belong here.
Contrary to what it claims, it doesn't define a reduce operator - it
merely defines a completely custom operator that happens to use
square brackets in its symbol.
Nor is is the functionality it implements related to reduce -
rather, it seems to re-implement (part of) the functionality of the
built-in `map` routine.